requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.11
django-cors-headers==4.3.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
idna==3.7
pillow==10.3.0
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.11.0
urllib3==2.2.1
python-dotenv
mysqlclient